Add call to translate api, translate title, plot and country
